pyproject.toml
[tool.poetry]
name = "electronic_sales_network"
version = "0.1.0"
description = ""
authors = ["Amir Gabduvaliev"]
readme = "README.md"
packages = [{include = "electronic_sales_network"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.1"
drf-spectacular = "^0.26.2"
djoser = "^2.2.0"
envparse = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
